What's new: 1. PDF View (pending) 2. Updated Dashboard as per new Stages 3. New Document upload main screen

+++ lib/models/summaries/dashboard_summary.dart
@@ -13,7 +13,7 @@ class DashBoardSummaryDTO {
     return DashBoardSummaryDTO(
       stage: json['stage'],
       count: json['count'],
-      loanAmount: json['loanAmount'],
+      loanAmount: json['loanAmount'].toDouble(),
     );
   }
